Refactoring and temporary commit before finishing the fix.
