+++ project/Release.scala
@@ -12,4 +12,5 @@ object Release {
       state
    }
 
+
 }
