+++ project/Sametest.scala
@@ -5,9 +5,6 @@ import Keys._
 
 // This code is adapted from scala.tools.ant.Same by Gilles Dubochet.
 object SameTest {
-  lazy val checkSame: TaskKey[Unit] = TaskKey("check-same-binaries", "checks whether or not the class files generated by scala are the same.")
-  lazy val checkSameLibrary: TaskKey[Unit] = TaskKey("check-same-lib-binaries", "checks whether or not the librayr class files generated by scala are the same.")
-  lazy val checkSameCompiler: TaskKey[Unit] = TaskKey("check-same-comp-binaries", "checks whether or not the compiler class files generated by scala are the same.")
 
   def checkSameBinaryProjects(lhs: Project, rhs: Project): Project.Initialize[Task[Unit]] =
     (classDirectory in Compile in lhs, classDirectory in Compile in rhs,
